protein-drug-binding/requirements.txt
torch>=2.0.0
torch-geometric>=2.3.0
e3nn>=0.5.0
rdkit>=2023.3.1
biopython>=1.81
py3Dmol>=2.0.0
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
scikit-learn>=1.3.0
jupyter>=1.0.0
wandb>=0.15.0
streamlit>=1.25.0